testApp/requirements.txt
Django==2.2.6
djangorestframework==3.8.2